My Portfolio/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0.2
django-ckeditor==6.2.0
django-js-asset==2.0.0
gunicorn==20.1.0
Pillow==9.0.1
psycopg2-binary==2.9.3
python-decouple==3.6
sqlparse==0.4.2
whitenoise==6.0.0
